Minor improvements and additions is utils
