+++ src/utils/classes/index.ts
@@ -1,0 +1,3 @@
+import { Time } from "./time";
+
+export { Time }
