+++ src/utils/functions/clamp.ts
@@ -1,0 +1,3 @@
+export function clamp(num: number, min: number, max: number): number {
+    return Math.min(max, Math.max(min, num));
+}
