+++ src/utils/functions/formatDate.ts
@@ -1,6 +1,6 @@
-import { DEFAULT_DATE, MONTHS } from "./consts";
+import { DEFAULT_DATE, MONTHS } from "../consts";
 
-export function formatDate(date: Date, withYear: boolean = false): string {
+export default function formatDate(date: Date, withYear: boolean = false): string {
   if(date === DEFAULT_DATE)
     return "";
