+++ src/utils/functions/index.ts
@@ -1,0 +1,4 @@
+import formatDate from "./formatDate";
+import { shadeDown } from "./shades";
+
+export { formatDate, shadeDown }
